Stub additional compiler modules
