+++ flux_lang/src/codegen/cranelift.rs
@@ -1,0 +1,6 @@
+use crate::ir::IrModule;
+
+pub fn emit_cranelift(_ir: &IrModule) -> Result<(), String> {
+    // TODO: implement Cranelift JIT backend
+    Ok(())
+}
