+++ flux_lang/src/codegen/llvm.rs
@@ -1,0 +1,6 @@
+use crate::ir::IrModule;
+
+pub fn emit_llvm(_ir: &IrModule) -> Result<(), String> {
+    // TODO: implement LLVM backend
+    Ok(())
+}
