+++ flux_lang/src/codegen/mod.rs
@@ -2,7 +2,11 @@
 
 use crate::ir::IrModule;
 
-pub fn emit(_ir: &IrModule) -> Result<(), String> {
-    // TODO: implement code generation
-    Ok(())
+pub mod cranelift;
+pub mod llvm;
+pub mod wasm;
+
+pub fn emit(ir: &IrModule) -> Result<(), String> {
+    // TODO: choose backend via options
+    llvm::emit_llvm(ir)
 }
