+++ flux_lang/src/codegen/wasm.rs
@@ -1,0 +1,6 @@
+use crate::ir::IrModule;
+
+pub fn emit_wasm(_ir: &IrModule) -> Result<(), String> {
+    // TODO: implement WASM backend
+    Ok(())
+}
