+++ flux_lang/src/ir/optimize.rs
@@ -1,0 +1,5 @@
+use super::IrModule;
+
+pub fn run_passes(_ir: &mut IrModule) {
+    // TODO: implement optimization passes
+}
