+++ flux_lang/src/semantic/env.rs
@@ -1,0 +1,10 @@
+//! Symbol table and type environments (placeholder)
+
+pub struct Env;
+
+#[allow(clippy::new_without_default)]
+impl Env {
+    pub fn new() -> Self {
+        Self
+    }
+}
