+++ flux_lang/src/semantic/mod.rs
@@ -1,0 +1,6 @@
+//! Semantic analysis and type checking stubs
+
+pub mod env;
+pub mod typecheck;
+
+pub use typecheck::check;
