+++ flux_lang/tests/typechecker_tests.rs
@@ -1,0 +1,6 @@
+use flux_lang::compile;
+
+#[test]
+fn typechecker_stub() {
+    assert!(compile("dummy").is_ok());
+}
